# a7_openapi.yaml
openapi: 3.0.0
info:
  version: "1.0"
  title: "LBAW Petopia Store Web API"
  description: "Web Resources Specification (A7) for Petopia Store"
servers:
  - url: http://lbaw.fe.up.pt
    description: Production server
...
paths:
...
  /login:
    get:
      operationId: R103
      summary: "R103: Log-in Form"
      description: "Provide log-in form. Access: PUB"
      tags:
        - "M01: Authentication and Personal Data"
      responses:
        "200":
          description: "Ok. Show log-in UI"

    post:
      operationId: R104
      summary: "R104: Log-in Action"
      description: "Processes the log-in form submission. Access: PUB"
      tags:
        - "M01: Authentication and Personal Data"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        "200":
          description: "Ok."
        "302":
          description: "Redirect after processing the log-in credentials."
          headers:
            Location:
              schema:
                type: string
              examples:
                302Success:
                  description: "Successful authentication. Redirect to home page."
                  value: "/home_page/{id}"
                302Error:
                  description: "Failed authentication. Redirect to log-in form."
                  value: "/login"
...
  /cart:
    get:
      operationId: R501
      summary: "R501: Display Cart"
      description: "Display the contents of the shopping cart. Access: USR, PBY"
      tags:
        - "M05: Cart, Wishlist and Reviews"
      responses:
        "200":
          description: "Ok. Show Cart UI"
        "404":
          description: "Not found. Cart is empty."
...